* common-src/amanda.h: Define __attribute__ for gcc >= 3. * common-src/packet.c (pkt_init): Accept NULL for fmt. * amandad-src/amandad.c, common-src/protocol.c,   common-src/security-util.c: Pass NULL to pkt_init. * amandad-src/amandad.c, changer-src/chg-scsi.c,   changer-src/scsi-linux.c, client-src/sendbackup.c,   common-src/amanda.h, common-src/conffile.c,   common-src/event.c, common-src/packet.c,   common-src/security-util.c, common-src/stream.c,   common-src/util.c, oldrecover-src/amrecover.c,   restore-src/restore.c, server-src/amcheck.c,   server-src/driver.c, server-src/driverio.c,   server-src/taper.c: Fix printf format error.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@144 a8d146d6-cc15-0410-8900-af154a0219e0

+++ changer-src/chg-scsi.c
@@ -200,7 +200,7 @@ dump_changer_struct(
   dbprintf (("\traw: %d\n",chg->emubarcode));
   if (chg->debuglevel != NULL)
      dbprintf(("debug level     : %s\n", chg->debuglevel));
-  dbprintf(("Tapes need sleep: %d seconds\n",chg->sleep));
+  dbprintf(("Tapes need sleep: %d seconds\n", (int)chg->sleep));
   dbprintf(("Cleancycles     : %d\n",chg->cleanmax));
   dbprintf(("Changerdevice   : %s\n",chg->device));
   if (chg->labelfile != NULL)
